+++ CHANGELOG.md
@@ -6,6 +6,7 @@ FlatLaf Change Log
 - Updated colors in "Flat Light" and "Flat IntelliJ" themes with colors from
   "IntelliJ Light Theme", which provides blue coloring that better match
   platform colors.
+- Tree: Support wide selection (enabled by default).
 - List and Tree: Paint cell focus indicator (black rectangle) only if more than
   one item is selected.
 - Fixed link color (in HTML text) and separator color in IntelliJ platform